# swagger.yaml
swagger: '2.0'
info:
  version: 0.0.1
  title: Henohenoapi
  description: >-
    The project aims to create a platform for managing exams, their submissions and the reviews related to the submissions
host: henoheno-app.herokuapp.com
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
...
securityDefinitions:
  Api_key:
    type: apiKey
    in: query
    name: token
security:
  - Api_key: []
paths:
...
  /users:
    get:
      summary: users list
      description: gets the list of all users registered on the platform (without password)
      tags:
        - Users_resource
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Users_in_response'
        '401':
          $ref: '#/responses/401_unauthorized'
        '500':
          $ref: '#/responses/500_internal_server_error'
    post:
      summary: user registration
      description: creates a new user based on the data posted
      tags:
        - Users_resource
      security:
        - {}
      parameters:
        - $ref: '#/parameters/User_in_post'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/User'
        '400':
          $ref: '#/responses/400_bad_request'
        '401':
          $ref: '#/responses/401_unauthorized'
        '500':
          $ref: '#/responses/500_internal_server_error'
    put:
      summary: user details update
      description: updates the information about an user(identified by his token)
      tags:
        - Users_resource
      parameters:
        - $ref: '#/parameters/User_in_put'
      responses:
        '200':
          $ref: '#/responses/200_ok_only'
        '401':
          $ref: '#/responses/401_unauthorized'
        '409':
          $ref: '#/responses/409_conflict'
        '500':
          $ref: '#/responses/500_internal_server_error'
  '/users/{user_id}':
    get:
      summary: access one specific user
      description: displays the data of an user given its id
      tags:
        - Users_resource
      parameters:
        - $ref: '#/parameters/User_id_in_path'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/User_in_response'
        '401':
          $ref: '#/responses/401_unauthorized'
        '404':
          $ref: '#/responses/404_not_found'
        '500':
          $ref: '#/responses/500_internal_server_error'
  '/users/{user_id}/exams':
    get:
      summary: get the exams of a student
      description: displays the exams created by an user or assigned to him
      tags:
        - Users_resource
      parameters:
        - $ref: '#/parameters/User_id_in_path'
        - $ref: '#/parameters/Selection_users_exams'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Exams'
        '401':
          $ref: '#/responses/401_unauthorized'
        '404':
          $ref: '#/responses/404_not_found'
        '500':
          $ref: '#/responses/500_internal_server_error'
...
  /tasks:
    get:
      summary: tasks list
      description: displays list of all tasks
      tags:
        - Tasks_resource
      parameters:
        - $ref: '#/parameters/Selection_tasks'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Tasks'
        '401':
          $ref: '#/responses/401_unauthorized'
        '400':
          $ref: '#/responses/400_bad_request'
        '500':
          $ref: '#/responses/500_internal_server_error'
    post:
      summary: create new task
      description: creates new task based on the posted data("options" can be empty if the task is an open-ended question - in this case the array of solutions can contain one single element which an example of a correct answer to the question)
      tags:
        - Tasks_resource
      parameters:
        - $ref: '#/parameters/Task_in_post'
      responses:
        '201':
          description: created
          schema:
            $ref: '#/definitions/Task'
        '204':
          $ref: '#/responses/204_no_content'
        '400':
          $ref: '#/responses/400_bad_request'
        '401':
          $ref: '#/responses/401_unauthorized'
        '500':
          $ref: '#/responses/500_internal_server_error'
  '/tasks/{task_id}':
    get:
      summary: access a specific task
      description: >-
        displays the informations of a task given its id(it won\'t contain the
        solution if you\'re not the owner)
      tags:
        - Tasks_resource
      parameters:
        - $ref: '#/parameters/Task_id_in_path'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Task'
        '401':
          $ref: '#/responses/401_unauthorized'
        '404':
          $ref: '#/responses/404_not_found'
        '500':
          $ref: '#/responses/500_internal_server_error'
    put:
      summary: update an existing task
      description: updates an existing task(can be done only by the owner of the task)
      tags:
        - Tasks_resource
      parameters:
        - $ref: '#/parameters/Task_id_in_path'
        - $ref: '#/parameters/Task_in_put'
      responses:
        '200':
          $ref: '#/responses/200_ok_only'
        '401':
          $ref: '#/responses/401_unauthorized'
        '404':
          $ref: '#/responses/404_not_found'
        '409':
          $ref: '#/responses/409_conflict'
        '500':
          $ref: '#/responses/500_internal_server_error'
    delete:
      summary: delete an existing task
      description: deletes an existing task
      tags:
        - Tasks_resource
      parameters:
        - $ref: '#/parameters/Task_id_in_path'
      responses:
        '204':
          $ref: '#/responses/204_no_content'
        '401':
          $ref: '#/responses/401_unauthorized'
        '404':
          $ref: '#/responses/404_not_found'
        '500':
          $ref: '#/responses/500_internal_server_error'
...
definitions:
  Error:
    properties:
      code:
        type: integer
      message:
        type: string
  User:
    type: object
    properties:
      id:
        type: integer
        description: the user's id
      email:
        type: string
        description: the user's email
      password:
        type: string
        format: password
        description: the user's password
      name:
        type: string
        description: the user's name
      surname:
        type: string
        description: the user's surname
...
  User_in_object:
    type: object
    properties:
      id:
        type: integer
      email:
        type: string
  Exam:
    type: object
    properties:
      id:
        type: integer
      owner:
        $ref: '#/definitions/User_in_object'
      title:
        type: string
      description:
        type: string
      taskset:
        type: array
        items:
          $ref: '#/definitions/Task_in_exam'
      group:
        $ref: '#/definitions/Group'
      final_deadline:
        type: string
        format: date
        description: date format "yyyy-mm-dd"
      review_deadline:
        type: string
        format: date
        description: 'date format "yyyy-mm-dd", review must be at least a day later'
  Exams:
    type: array
    items:
      $ref: '#/definitions/Exam'
  Task:
    type: object
    properties:
      id:
        type: integer
      owner:
        $ref: '#/definitions/User_in_object'
      task_type:
        type: string
      subject:
        type: string
      title:
        type: string
      description:
        type: string
      options:
        type: array
        items:
          type: string
      solutions:
        type: array
        items:
          type: string
  Tasks:
    type: array
    items:
      $ref: '#/definitions/Task'
  Task_in_exam:
    type: object
    properties:
      id:
        type: integer
      description:
        type: string
  Group:
    type: object
    properties:
      id:
        type: integer
      owner:
        $ref: '#/definitions/User_in_object'
      name:
        type: string
      description:
        type: string
      members:
        type: array
        items:
          $ref: '#/definitions/User_in_object'
...
  User_in_response:
    type: object
    properties:
      id:
        type: integer
        description: the user's id
      email:
        type: string
        description: the user's email
      name:
        type: string
        description: the user's name
      surname:
        type: string
        description: the user's surname
  Users_in_response:
    type: array
    items:
      $ref: '#/definitions/User_in_response'
  User_for_post:
    type: object
    properties:
      email:
        type: string
        description: the user's email
      password:
        type: string
        format: password
        description: the user's password
      name:
        type: string
        description: the user's name
      surname:
        type: string
        description: the user's surname
...
  Task_for_post:
    type: object
    properties:
      task_type:
        type: string
      subject:
        type: string
      title:
        type: string
      description:
        type: string
      options:
        type: array
        description: 'arrat format "aaa,bbb,ccc," deve avere almeno 2 elementi'
        items:
          type: string
      solutions:
        type: array
        items:
          type: string
...
parameters:
...
  Selection_users_exams:
    in: query
    name: select
    description: select a particular type of exams
    required: true
    type: string
    enum:
      - created
      - assigned
...
  Selection_tasks:
    in: query
    name: select
    description: select a particular type of tasks
    required: true
    type: string
    enum:
      - created
      - all
...
  User_in_post:
    name: user
    description: the object user
    in: body
    schema:
      $ref: '#/definitions/User_for_post'
      required:
        - email
        - password
        - name
        - surname
  User_in_put:
    name: user
    description: the object user
    in: body
    schema:
      $ref: '#/definitions/User_for_post'
      required:
        - email
        - password
        - name
        - surname
...
  User_id_in_path:
    name: user_id
    description: the user's id
    in: path
    type: integer
    required: true
...
  Task_in_post:
    name: task
    description: the object task
    in: body
    schema:
      $ref: '#/definitions/Task_for_post'
      required:
        - task_type
        - subject
        - title
        - description
        - options
        - solutions
  Task_in_put:
    name: task
    description: the object task
    in: body
    schema:
      $ref: '#/definitions/Task_for_post'
      required:
        - task_type
        - subject
        - title
        - description
        - options
        - solutions
  Task_id_in_path:
    name: task_id
    description: id task
    in: path
    type: integer
    required: true
...
responses:
...
  200_ok_only:
    description: OK
...
  204_no_content:
    description: No Content
  400_bad_request:
    description: Bad Request
    schema:
      $ref: '#/definitions/Error'
  401_unauthorized:
    description: 'Unauthorized, missing or invalid API Key'
    schema:
      $ref: '#/definitions/Error'
...
  404_not_found:
    description: Not Found
    schema:
      $ref: '#/definitions/Error'
...
  409_conflict:
    description: Conflict
    schema:
      $ref: '#/definitions/Error'
  500_internal_server_error:
    description: Internal Server Error
    schema:
      $ref: '#/definitions/Error'